apps/subsquid-silo-tests/data-ingestion/requirements.txt
# ========================================
# Subsquid Silo Tests - Python Dependencies
# ========================================
# Async & HTTP
httpx==0.24.1
aiohttp==3.9.1
websockets==12.0
asyncio-contextmanager==1.0.1

# Database
asyncpg==0.29.0
psycopg==3.2.11
psycopg-binary==3.2.11
sqlalchemy==2.0.23
alembic==1.12.1

# Redis
redis==5.0.0
aioredis==2.0.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# AI Categorization
openai==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
responses==0.24.1

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Logging & Monitoring
structlog==23.2.0

# (Optional) DipDup for on-chain indexing
# dipdup==7.0.0  # Will be installed separately with dipdup init
